feat: add user context provider and axios instance for API interactions
- Added UserContext to manage user state and provide user data throughout the application.
- Implemented axiosInstance with request interceptor for handling token expiration and refreshing tokens.
- Updated models to include User and UserResponse interfaces.
- Added new dependencies: axios, bcrypt, jsonwebtoken, and nodemailer.
