+++ app/(auth)/login/page.tsx
@@ -1,0 +1,119 @@
+"use client";
+
+import { useEffect, useState } from "react";
+import { useRouter } from "next/navigation";
+import { Button } from "@/components/ui/button";
+import { Input } from "@/components/ui/input";
+import { Eye, EyeOff, LogIn } from "lucide-react";
+import Link from "next/link";
+import axios from '../../../utils/axiosInstance'
+import { useUser } from "@/providers/usercontext";
+
+export default function LoginPage() {
+  const {user} = useUser();
+  const router = useRouter();
+  const [showPassword, setShowPassword] = useState(false);
+  const [email, setEmail] = useState("");
+  const [password, setPassword] = useState("");
+  const [error, setError] = useState<string | null>(null);
+  const [loading, setLoading] = useState(false);
+
+  useEffect(() => {
+    if(user){
+      router.push('/')
+    }
+  }, [user, router])
+
+  const handleLogin = async (e: React.FormEvent) => {
+    e.preventDefault();
+    setError(null);
+    setLoading(true);
+
+    try {
+      const res = await axios.post("/auth/login", { email, password });
+      const { accessToken, refreshToken } = res.data;
+
+      localStorage.setItem("accessToken", accessToken);
+      localStorage.setItem("refreshToken", refreshToken);
+
+      router.push("/");
+    } catch (err: any) {
+      setError(err.response?.data?.error || "Login failed. Try again.");
+    } finally {
+      setLoading(false);
+    }
+  };
+
+  return (
+    <div className="min-h-screen flex items-center justify-center bg-gradient-to-br from-slate-50 via-emerald-50/20 to-slate-100 p-4">
+      <div className="w-full max-w-md bg-white rounded-2xl shadow-sm border border-slate-200/60 p-8 space-y-6">
+        <div className="text-center">
+          <h1 className="text-2xl font-bold text-slate-900 mb-2">Login</h1>
+          <p className="text-slate-600">Access your ColdStore Monitor dashboard</p>
+        </div>
+
+        <form className="space-y-4" onSubmit={handleLogin}>
+          <div>
+            <label className="block text-sm font-medium text-slate-700 mb-1">
+              Email
+            </label>
+            <Input
+              type="email"
+              placeholder="you@example.com"
+              value={email}
+              onChange={(e) => setEmail(e.target.value)}
+              required
+            />
+          </div>
+
+          <div>
+            <label className="block text-sm font-medium text-slate-700 mb-1">
+              Password
+            </label>
+            <div className="relative">
+              <Input
+                type={showPassword ? "text" : "password"}
+                placeholder="••••••••"
+                value={password}
+                onChange={(e) => setPassword(e.target.value)}
+                required
+              />
+              <button
+                type="button"
+                onClick={() => setShowPassword(!showPassword)}
+                className="absolute inset-y-0 right-3 flex items-center text-slate-500 hover:text-slate-700"
+              >
+                {showPassword ? (
+                  <EyeOff className="h-4 w-4" />
+                ) : (
+                  <Eye className="h-4 w-4" />
+                )}
+              </button>
+            </div>
+          </div>
+
+          {error && <p className="text-sm text-red-600">{error}</p>}
+
+          <Button
+            type="submit"
+            className="w-full bg-emerald-600 hover:bg-emerald-700 text-white"
+            disabled={loading}
+          >
+            <LogIn className="w-4 h-4 mr-2" />
+            {loading ? "Logging in..." : "Login"}
+          </Button>
+        </form>
+
+        <p className="text-sm text-slate-600 text-center">
+          Don&apos;t have an account?{" "}
+          <Link
+            href="/register"
+            className="text-emerald-600 hover:underline font-medium"
+          >
+            Sign up
+          </Link>
+        </p>
+      </div>
+    </div>
+  );
+}
