+++ app/api/auth/login/route.tsx
@@ -1,0 +1,71 @@
+import { NextRequest, NextResponse } from "next/server";
+import bcrypt from "bcrypt";
+import jwt from "jsonwebtoken";
+import { query } from "@/lib/db";
+
+const JWT_SECRET = process.env.JWT_SECRET || "dev-secret";
+
+export async function POST(req: NextRequest) {
+  try {
+    const { email, password } = await req.json();
+
+    if (!email || !password) {
+      return NextResponse.json(
+        { error: "Email and password are required." },
+        { status: 400 }
+      );
+    }
+
+    // Fetch user from DB
+    const rows = await query<{
+      id: string;
+      email: string;
+      password: string;
+    }>(
+      `
+      SELECT id, email, password
+      FROM "user"
+      WHERE email = $1
+      LIMIT 1
+      `,
+      [email]
+    );
+
+    if (!rows.length) {
+      return NextResponse.json({ error: "Invalid credentials." }, { status: 401 });
+    }
+
+    const user = rows[0];
+
+    // Compare password
+    const isMatch = await bcrypt.compare(password, user.password);
+    if (!isMatch) {
+      return NextResponse.json({ error: "Invalid credentials." }, { status: 401 });
+    }
+
+    // Generate tokens
+    const accessToken = jwt.sign(
+      { userId: user.id, email: user.email },
+      JWT_SECRET,
+      { expiresIn: "1d" }
+    );
+
+    const refreshToken = jwt.sign(
+      { userId: user.id, email: user.email },
+      JWT_SECRET,
+      { expiresIn: "7d" }
+    );
+
+    return NextResponse.json({
+      accessToken,
+      refreshToken,
+      user: { id: user.id, email: user.email },
+    });
+  } catch (error: any) {
+    console.error("Login error:", error);
+    return NextResponse.json(
+      { error: "Internal server error", detail: error.message },
+      { status: 500 }
+    );
+  }
+}
