+++ app/api/logs/route.ts
@@ -1,9 +1,46 @@
 import { NextRequest, NextResponse } from "next/server";
 import { query } from "@/lib/db";
 import { toLogFormat } from "@/lib/utilities"; // <-- use your helper
+import jwt from 'jsonwebtoken';
+async function getUserFromRequest(req: NextRequest) {
+  try {
+    const authHeader = req.headers.get("authorization");
+    if (!authHeader) return null;
+
+    const token = authHeader.replace("Bearer ", "");
+    const decoded: any = jwt.verify(token, process.env.JWT_SECRET as string);
+    const rows = await query<{
+      id: string;
+      email: string;
+      password: string;
+    }>(
+      `
+      SELECT id, email, password
+      FROM "user"
+      WHERE id = $1
+      LIMIT 1
+      `,
+      [decoded.userId.userId.userId.userId]
+    );
+
+    if (!rows.length) {
+      return NextResponse.json({ error: "Invalid credentials." }, { status: 401 });
+    }
+
+    const user = rows[0];
+    return user;
+  } catch {
+    return null;
+  }
+}
 
 export async function GET(req: NextRequest) {
   try {
+    const user = await getUserFromRequest(req);
+    if (!user) {
+      return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
+    } 
+
     const { searchParams } = new URL(req.url);
 
     const start = searchParams.get("start");
