+++ app/api/zones/[id]/route.ts
@@ -1,12 +1,49 @@
 import { NextRequest, NextResponse } from "next/server";
 import { query } from "@/lib/db";
+import jwt from 'jsonwebtoken';
 
+async function getUserFromRequest(req: NextRequest) {
+  try {
+    const authHeader = req.headers.get("authorization");
+    if (!authHeader) return null;
+
+    const token = authHeader.replace("Bearer ", "");
+    const decoded: any = jwt.verify(token, process.env.JWT_SECRET as string);
+    const rows = await query<{
+      id: string;
+      email: string;
+      password: string;
+    }>(
+      `
+      SELECT id, email, password
+      FROM "user"
+      WHERE id = $1
+      LIMIT 1
+      `,
+      [decoded.userId]
+    );
+
+    if (!rows.length) {
+      return NextResponse.json({ error: "Invalid credentials." }, { status: 401 });
+    }
+
+    const user = rows[0];
+    return user;
+  } catch {
+    return null;
+  }
+}
 export async function GET(
   req: NextRequest,
   context: { params: Promise<{ id: string }> }
 ) {
   const { id } = await context.params;
   try {
+    const user = await getUserFromRequest(req);
+    if (!user) {
+      return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
+    } 
+
     const rows = await query<any>(
       `SELECT id, name, min_c, max_c, humidity_min, humidity_max 
        FROM zones WHERE id = $1`,
@@ -33,6 +70,10 @@ export async function PUT(
 ) {
   const { id } = await context.params;
   try {
+    const user = await getUserFromRequest(req);
+    if (!user) {
+      return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
+    }
     const body = await req.json();
     const { name, min_c, max_c, humidity_min, humidity_max } = body;
 
@@ -64,6 +105,10 @@ export async function DELETE(
 ) {
   const { id } = await context.params;
   try {
+    const user = await getUserFromRequest(req);
+    if (!user) {
+      return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
+    }
     const row = await query<any>(
       `DELETE FROM zones WHERE id = $1 RETURNING id`,
       [id]
